LU-5577 changelog: fix comparison between signed and unsigned
Change type of changelog_*{namelen,size}() to size_t.
Fixed string specifier for unsigned types.

Reviewed-on: http://review.whamcloud.com/12474
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/lustre/lustre_user.h
@@ -866,9 +866,9 @@ struct changelog_ext_jobid {
 };
 
 
-static inline unsigned int changelog_rec_offset(enum changelog_rec_flags crf)
+static inline size_t changelog_rec_offset(enum changelog_rec_flags crf)
 {
-	unsigned int size = sizeof(struct changelog_rec);
+	size_t size = sizeof(struct changelog_rec);
 
 	if (crf & CLF_RENAME)
 		size += sizeof(struct changelog_ext_rename);
@@ -879,12 +879,12 @@ static inline unsigned int changelog_rec_offset(enum changelog_rec_flags crf)
 	return size;
 }
 
-static inline unsigned int changelog_rec_size(struct changelog_rec *rec)
+static inline size_t changelog_rec_size(struct changelog_rec *rec)
 {
 	return changelog_rec_offset(rec->cr_flags);
 }
 
-static inline unsigned int changelog_rec_varsize(struct changelog_rec *rec)
+static inline size_t changelog_rec_varsize(struct changelog_rec *rec)
 {
 	return changelog_rec_size(rec) - sizeof(*rec) + rec->cr_namelen;
 }
@@ -916,7 +916,7 @@ static inline char *changelog_rec_name(struct changelog_rec *rec)
 						  CLF_SUPPORTED);
 }
 
-static inline int changelog_rec_snamelen(struct changelog_rec *rec)
+static inline size_t changelog_rec_snamelen(struct changelog_rec *rec)
 {
 	return rec->cr_namelen - strlen(changelog_rec_name(rec)) - 1;
 }
